requirements.txt
asgiref==3.5.2
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
django-debug-toolbar==3.6.0
django-rest-framework==0.1.0
django-split-settings==1.1.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
flake8==5.0.4
h11==0.13.0
mccabe==0.7.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
uvicorn==0.18.3
whitenoise==6.2.0
